+++ packages/bluesky/lib/src/services/entities/list_view_basic.dart
@@ -5,6 +5,7 @@
 // ignore_for_file: invalid_annotation_target
 
 // 📦 Package imports:
+import 'package:atproto/atproto.dart';
 import 'package:atproto_core/atproto_core.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
 
@@ -32,6 +33,7 @@ class ListViewBasic with _$ListViewBasic {
     required String cid,
     required String name,
     String? avatar,
+    List<Label>? labels,
     @Default(defaultListViewer) ListViewer viewer,
     required DateTime indexedAt,
   }) = _ListViewBasic;