+++ packages/bluesky/lib/src/services/entities/list_view_basic.g.dart
@@ -22,6 +22,12 @@ _$ListViewBasicImpl _$$ListViewBasicImplFromJson(Map json) => $checkedCreate(
           cid: $checkedConvert('cid', (v) => v as String),
           name: $checkedConvert('name', (v) => v as String),
           avatar: $checkedConvert('avatar', (v) => v as String?),
+          labels: $checkedConvert(
+              'labels',
+              (v) => (v as List<dynamic>?)
+                  ?.map((e) =>
+                      Label.fromJson(Map<String, Object?>.from(e as Map)))
+                  .toList()),
           viewer: $checkedConvert(
               'viewer',
               (v) => v == null
@@ -51,6 +57,7 @@ Map<String, dynamic> _$$ListViewBasicImplToJson(_$ListViewBasicImpl instance) {
   }
 
   writeNotNull('avatar', instance.avatar);
+  writeNotNull('labels', instance.labels?.map((e) => e.toJson()).toList());
   val['viewer'] = instance.viewer.toJson();
   val['indexedAt'] = instance.indexedAt.toIso8601String();
   return val;